# src/main/resources/static/openapi-custom.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  description: API of the "School" task
  contact:
    name: Zhadaev Alexander
    url: www.zhadaev.ru
    email: zhadaevs@yandex.ru
  version: 1.0-SNAPSHOT
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /api/groups/{id}:
    summary: Represents a group
    description: >
      This resource represents an individual group in the system.
      Each group is identified by a numeric `id`.
    get:
      summary: Get a group by id
      description: >
        This method allows you to get a school group by its id.
      tags:
        - group-controller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The group id
          example: 46fa82ce-4e6d-45ae-a4e4-914971f1eb4f
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
              example:
                id: 46fa82ce-4e6d-45ae-a4e4-914971f1eb4f
                name: YT-80
        '404':
          description: Group not found by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
              example:
                timestamp: "2022-11-17T17:47:01.030+00:00"
                status: Not Found
                message: Group not found by id
    put:
      summary: Edit a group
      description: >
        This method updates the group with the specified id. The values of the group properties are replaced with the values from the request body. Unspecified values are replaced with default values.
      tags:
        - group-controller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The group id
          example: 46fa82ce-4e6d-45ae-a4e4-914971f1eb4f
      requestBody:
        description: >
          You must specify the group parameters.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupDto'
            example:
              name: YT-80
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
              example:
                id: 46fa82ce-4e6d-45ae-a4e4-914971f1eb4f
                name: YT-80
        '404':
          description: Group replace error. Group not found by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
              example:
                timestamp: "2022-11-17T17:47:01.030+00:00"
                status: Not Found
                message: Group replace error. Group not found by id
        '400':
          description: >
            Validation error. The group name must not be null and must contain at least one non-whitespace character.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              example:
                timestamp: "2022-11-17T19:07:45.652+00:0"
                status: Bad Request
                violations:
                  - message: The group name must not be null and must contain at least one non-whitespace character
                    propertyPath: groupDto
                    invalidValue: "  "
    delete:
      summary: Delete a group
      description: >
        This method removes the group with the specified id.
      tags:
        - group-controller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The group id
          example: 46fa82ce-4e6d-45ae-a4e4-914971f1eb4f
      responses:
        '200':
          description: Successful operation
        '404':
          description: Group delete error. Group not found by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
              example:
                timestamp: "2022-11-17T17:47:01.030+00:00"
                status: Not Found
                message: Group delete error. Group not found by id
    patch:
      summary: Edit a group
      description: >
        This method updates the group with the specified id. The values of the group properties are replaced with the values from the request body. Unspecified values are not changed.
      tags:
        - group-controller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The group id
          example: 46fa82ce-4e6d-45ae-a4e4-914971f1eb4f
      requestBody:
        description: >
          You must specify the group parameters.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupDto'
            example:
              name: YT-80
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
              example:
                id: 46fa82ce-4e6d-45ae-a4e4-914971f1eb4f
                name: YT-80
        '404':
          description: Group update error. Group not found by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
              example:
                timestamp: "2022-11-17T17:47:01.030+00:00"
                status: Not Found
                message: Group update error. Group not found by id
        '400':
          description: >
            Validation error. The group name must not be null and must contain at least one non-whitespace character.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              example:
                timestamp: "2022-11-17T19:07:45.652+00:0"
                status: Bad Request
                violations:
                  - message: The group name must not be null and must contain at least one non-whitespace character
                    propertyPath: groupDto
                    invalidValue: "  "
...
  /api/groups:
    summary: Represents a groups
    description: >
      This resource represents an all groups in the system.
    get:
      summary: Get all groups
      description: >
        This method allows you to get a list of all school groups. If the filter parameter is set, the method returns a list of filtered groups. In this case, the method returns groups in which the number of students is less than or equal to the specified.
      tags:
        - group-controller
      parameters:
        - name: numberStudents
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
          description: Number of students
          example: 15
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
          description: Page number, size, sorting
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupDto'
              example:
                - id: 408a9358-c6b1-4b36-8912-7bbd4803f1b1
                  name: BA-51
                - id: d73d111e-7159-48e5-9247-e050db9e0437
                  name: OU-60
                - id: 10562d54-0acc-4fbf-baba-98c0aa77a900
                  name: SV-51
        '400':
          description: >
            Validation error. The number of students must be greater than or equal to zero.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              example:
                timestamp: "2022-11-17T19:07:45.652+00:0"
                status: Bad Request
                violations:
                  - message: The number of students must be greater than or equal to zero
                    propertyPath: findAll.numberStudents
                    invalidValue: -2
    post:
      summary: Add a new group
      description: >
        This method adds a new group to the school.
      tags:
        - group-controller
      requestBody:
        description: >
          You must specify the group parameters.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupDto'
            example:
              name: SV-51
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDto'
              example:
                id: c20bf536-996e-45cb-8f8d-b57f7b9a5998
                name: SV-51
        '400':
          description: >
            Validation error. The group name must not be null and must contain at least one non-whitespace character.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              example:
                timestamp: "2022-11-17T19:07:45.652+00:0"
                status: Bad Request
                violations:
                  - message: The group name must not be null and must contain at least one non-whitespace character
                    propertyPath: groupDto
                    invalidValue: "  "
    delete:
      summary: Delete all groups
      description: >
        This method removes all groups.
      tags:
        - group-controller
      responses:
        '200':
          description: Successful operation
...
components:
  schemas:
...
    GroupDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    CustomError:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: string
        message:
          type: string
    Violation:
      type: object
      properties:
        message:
          type: string
        propertyPath:
          type: string
        invalidValue:
          type: object
    ValidationError:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: string
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string